Refactor: 비슷한 게임 로직을 묶은 커스텀 훅 추가
